norminette good chackmap & libft

+++ srcs/dependency/libft/ft_strnstr.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, int len)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (*needle == '\0')
